// apps/web/src/lib/api-client/models/categoryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryDto {
  /** Identificador único da categoria */
  id: string;
  /** Nome da categoria */
  name: string;
  /** Identificador do usuário proprietário */
  ownerId: string;
  /** Data de criação */
  createdAt: string;
  /** Data de última atualização */
  updatedAt: string;
}